Remove unnecessary service start
The Dockerfile instructs the creation of a Docker image, so the concept of a running service in the Dockerfile doesn't make sense. The services would need to be tied to the CMD or ENTRYPOINT of the image, or run in a script within the container. The latter is what is happening already, so calling services in the Dockerfile is superfluous. Whether or not any of that is actually true, the image works fine without it.