Add Svelte action to insert an element
This commits creates a new file in the components lib for actions. As
there is only one generic action so far, the file is intentionally
generic. I expect that this file will contain a few more actions however
if we eventually have a lot of them it will be worth splitting into
different files or even have a directory for them.