Update Tabview component to use Svelte actions
I have added comments noting where the structure and naming scheme of
the tabview diverge from Wikidot's. Given that the old structure was so
dependent on Yahoo UI, it did not make sense to retain these class
names. Under normal circumstances I would have attempted to reproduce
Wikidot's naming scheme where possible, however, in this case the actual
structure remains intact.

A simple 'Clickable' component has been added that is simply an anchor
with a no-op href. I extracted this out to a component to quarantine its
lint errors. This component will become decreasingly necessary as we
phase out no-op anchors in favour of buttons.

Throughout the Tabview components, 'tabNumber' has been used to refer to
the index of each tab, as opposed to 'tabIndex' which could be ambiguous
with the HTML feature.