Move the Tabview component back into the src root
I had moved the Tabview component into the stubs dir to indicate that it
consumes a stub component, but as it is not itself a stub component,
this doesn't make much sense. I am not sure whether or not there needs
to be a way to distinguish them.

Once components are split across multiple dirs, and especially if we end
up having multiple components with the same name, we will need a way for
a given Svelte component to programmatically declare if it requires a
stub specification, and if so, which one. Later, we will be able to use
this information to extract the `query` from the spec to know how to
inject the Svelte component into the page.