Move query to spec root
I've seen conflicting information for whether the '$ref' property of a
JSON schema can coexist with other properties. I've been assuming that
it can, and WebStorm's built-in schema validation seems to support that.

However, some information I've seen asserts that using $ref alongside
other properties will result in those other properties simply being
ignored. This seems to be the case for jsonschema, the NPM package I'm
using for schema validation in tests, resulting in inaccurate
validation.

I've decided to err on the side of caution and make sure $ref is never
used in tandem with other properties, which seems to be working fine for
both jsonschema and WebStorm so far. I've allowed $ref to coexist with
no-op properties, though, such as 'description' (which is useful for
validation in editors) and duplicate properties that enhance readability
but don't do anything by virtue of being duplicate (like 'type: object'
when referencing '#/$defs/element').

This commit moves 'query' to the top level of the spec instead of being
an addition to 'root', which gives it its own namespace instead of
trying to extend the element definition.