Create initial stub validation script
This commits adds a script that, given a HTMLElement object, verifies
that it matches a specification (so long as that specification matches
the stubs.schema.yaml schema).

It's only partially complete. Currently it assumes that only a single
HTMLElement will be passed to it, but that's not always the case - the
schema allows for more than one element to be passed at a time,
including the root element, on account of the 'repeats' option. To make
better use of recursion, at all steps it should not assume this, and
should accept an array of elements.