Import YAML files as string
Tests for the stub specifications require reading the YAML files that
contain them. For this I would have used fs.readFile or similar to read
the file, as I assume that these tests will only ever be run on Node
(it would definitely be a mistake to ever run the stub validation script
in the browser). However, because the tests are compiled to
/tests-dist/, this breaks all the relative file URIs. Instead, esbuild
has been instructed to resolve YAML imports as strings, so they will now
be collected at build time.