Remove nested nabs scripts
Nested nabs script declarations produce scripts starting with 'npm run',
which doesn't match what is currently in the package.json (which start
with 'npm'). This commit replaces any nesting with manual script
concatenation to ensure that all commands start with 'pnpm'.

This commit also replaces any nabs scripts combined with '&&' to YAML
sequences (except for the test runner, which will be changed in the next
commit), which will be more readable as more tasks are added to them.