Check partials in rule execution.
This way if it fails, it properly resets the parser state
using the normal flow, and avoids top-level errors.

The problem here was that checking for partials *in the success
branch* of the consume() call (specifically the try_consume() for
the rule), is not expected, which yields the top-level errors
we observed prior, as well as not benefitting from the reset
functionality failing rules normally provide.