Add allow-plugins field.
Composer is making a change where the current default value
for this setting (null, meaning all plugins can be used)
will be set to {}. The allow-plugins configuration field is
an explicit list of all plugins which are permitted, in order
to use a plugin, it must be listed here.

Because composer wants people to make this composer.json change
before they roll out the breaking change (changing the default
behavior), they have a script which adds it to your composer.json
for you. However, like any time you run PHP, it loads all the
environment and other autoload stuff, which for us fails because
not all of the source is in the image yet (and besides we have a bunch
of extensions and dependencies that need to exist -- dependencies which
aren't present yet because this occurs while we're installing
dependencies!)