Replace String::replace_range() with str.replace().
The previous method is required for in-place modification, but because
we do not modify the original, we can just use the standard library
method str::replace() instead, which does not have the infinite loop
issue. Because find() / replace_range() is operating on an unchanged
string, the while condition is always true, so it always loops.