Shift towards alternate structure for storing deleted pages.
This empty commit is being used to document some thoughts about
how deleted pages are currently stored, and how I will be changing
the paradigm to better represent the data.

The "page" table already has the "deleted_at" column, which memorializes
whether the given page is currently deleted. I do not think this needs
to change.

"Revision delete" as a system is not the same as in MediaWiki, instead
the "hidden" column in the "page_revision" table allows hiding relevant
data. Instead of a "deleted revision" as a separate concept, we just
consider one effectively deleted if whatever relevant data points from
the revision have been hidden.

So considering the current way that page deletion is handled:

There is the ill-defined "metadata" column, which currently has the only
key, "deleted" (bool). This cannot be a loose key-value store like for custom
metadata (see https://scuttle.atlassian.net/browse/WJ-129) since some
fields like "deleted" need to be handled particularly at the platform
level.

Instead, because there is no clear use for the "metadata" columnyet, and
we are not working on custom data fields for a while, I am moving to
remove that column from pages, and add a different concept to
"page_revision".

The "revision_type" column in "page_revision" will be an enum value
(default 'regular'). This adds the values 'delete' and 'undelete' to
represent tombstone and resurrection revisions accordingly. This also
allows us to add the 'create' revision type to reflect new pages.

This corrects for an issue with the JSON metadata system, where it is
not clear if "deleted" is meant to reflect a current state (i.e. the
page is currently extant / not deleted), or an event (i.e. in this
revision, the page is being deleted).