Replace RevisionTasks::determine() logic.
Since we now have "changes" look at what's different from previous to
proposed, we can just consult that when building our task list.

This also resolves the TODO comment calling for the change logic to be
consolidated. (It is now, in RevisionService where "changes" is created,
and outdate / update tasks are performed).