Move vote queries inside Scorer impls.
This enhances performance, since instead of a relatively-expensive
GROUP BY aggregate query, it can query the vote result directly.
The database can simply sum up the votes etc itself, or for the
NullScorer, it doesn't need to use the database at all.

The VoteMap collection system is still present as a utility in case
it's helpful, but it's not mandatory.

This also removes the VoteMap auto-conversion to float.