Adjust build workflow activations to apply on PR.
This way, builds are always run, even if the changes come from outside
the repository (i.e. a fork -> PR).

This also has the consequence of always running these workflows when you
merge to develop or prod, but I consider this acceptable because we
should be checking everything before deploy anyways.