Just construct the ALTER SEQUENCE queries manually...
Really bummed that my attempts to use parameters wasn't working, it's
really not clear why it compiles but it can't substitute my parameters
for some reason. Oh well, this is known as static / hardcoded so it's
fine, and the seeder is not able to be called by the user anyways.