Split auth password/mfa, restructure for constant time.
Previously this would require a user to be pulled before authentication,
which could open the system to timing accounts as it exposes whether
a certain user is in the system or not.

Now, even if the name / email is not present, it will pull incorrect
user data and then fail on that, so the time taken is more consistent.