Remove session invalidation TODO on password change.
Wikidot added this as a weird "invalidate all other sessions"
implementation, but the notable difference is that Wikidot has
sessions which last an indefinite period of time, where invalidation
from account capture is a big deal. With a 30 minute token expiry time
and the ability to invalidate all other sessions, Wikijump does not
have this issue, and invalidating during password changing is not really
necessary.