Don't short-circuit user check if no alias.
The prior comment explained the issue, but I'm not sure why this was the case.
After all, it's not like we create duplicate user_alias entries for each user.slug,
in fact we want to hold the invariant where any given (live) user slug exists only
once between the user and user_alias tables.