Implement after-call future pattern for solving alias update issue.
To describe this a bit, there's an issue we run into when it comes to
adding an alias for the old slug.

Say we're renaming something from 'foo' -> 'bar' and want to leave an
alias behind. Before the user's slug is actually changed to 'bar', you
cannot actually create the user redirect until it's changed.

But update methods rely on doing "setup" for each and then the actual
UPDATE query at the end. So we're at a dilemma.

If we need to create an alias, we instead package that into a future
and return that, which the caller then invokes after the update has
gone through in the transaction. This fixes our dependency issue.

Though note that this was previously solved through the user of the
method "create_no_verify()", which only properly checked for slug
collisions manually ("verify()" method) after the UPDATE occurred.
However when the exists() methods on the services were changed to
actually check aliases too, this made it so that create_no_verify()
really wasn't much different than create() when it comes to what
we care about here, which is the ordering.

So an alternate solution is to fix that by not checking for existence
until afterwards and restoring "create_no_verify()" to doing what it
was originally intended to.