Get rid of after-call future pattern in favor of manual verification.
Because the issue (as described in e2318a2e9) is that the slug
checks in AliasService::create() were effectively verifying anyways,
we instead make that check wrapped in an if-statement, and by turning
that off ("create_no_verify()") we actually create without verifying.

Additionally we don't need to store the "needs_verify" flag since we
can determine that simply by looking at whether the entity's slug was
updated, in which case we check anyways.