Use new health check.
See https://github.com/minio/minio/issues/18373#issuecomment-1793010380.
Breaking change in upstream minio container.

Instead of replacing the healthcheck everywhere, we're moving it to a
script instead so that one place can be called, and if the healthcheck
is ever updated in the future, we only need to do so in one place.